nomad: don't break on allocation pending on a client
We regularly got errors such as:

    Check failed to run: undefined method `each' for nil:NilClass,
    ["/opt/sensu/embedded/lib/ruby/gems/2.3.0/gems/sensu-plugins-edgelab-1.6.1/bin/check-nomad-jobs.rb:126:in `block in check_allocations'",
    "/opt/sensu/embedded/lib/ruby/gems/2.3.0/gems/sensu-plugins-edgelab-1.6.1/bin/check-nomad-jobs.rb:121:in `each'",
    "/opt/sensu/embedded/lib/ruby/gems/2.3.0/gems/sensu-plugins-edgelab-1.6.1/bin/check-nomad-jobs.rb:121:in `check_allocations'",
    "/opt/sensu/embedded/lib/ruby/gems/2.3.0/gems/sensu-plugins-edgelab-1.6.1/bin/check-nomad-jobs.rb:184:in `block in run'",
    "/opt/sensu/embedded/lib/ruby/gems/2.3.0/gems/sensu-plugins-edgelab-1.6.1/bin/check-nomad-jobs.rb:182:in `each'",
    "/opt/sensu/embedded/lib/ruby/gems/2.3.0/gems/sensu-plugins-edgelab-1.6.1/bin/check-nomad-jobs.rb:182:in `run'",
    "/opt/sensu/embedded/lib/ruby/gems/2.3.0/gems/sensu-plugin-1.4.4/lib/sensu-plugin/cli.rb:58:in `block in '"]

This is due to some allocations having a TaskStates set to nil and the
ClientStatus set to pending, I suppose while the allocation as been
placed on a client and the client is downloading a Docker image (for
example).

We don't break on this, and we try also to see if there are allocations
which have been pending for too long on a client.